refactor(web): move game info to its own component
